thunder-react: move spec to separate file
