+++ client/src/spec.ts
@@ -1,0 +1,17 @@
+export interface MutationSpec<
+  Result extends object,
+  Input extends object | undefined = undefined
+> {
+  query: string;
+  result?: Result;
+  variables?: Input;
+}
+
+export interface QuerySpec<
+  Result extends object,
+  Input extends object | undefined = undefined
+> {
+  query: string;
+  result?: Result;
+  variables?: Input;
+}
